Moved search to lex
