+++ pages/lex/+Page.module.sass
@@ -12,3 +12,29 @@
     .stat:first-child
         border-left: 1px solid var(--text-color)
         
+
+.search-container
+
+  .search-results
+    background-color: var(--panel-background-color)
+
+    .category
+      text-align: left
+      background-color: var(--accent-secondary-color)
+      padding: 5px 10px
+      margin: 0
+      border-top: solid var(--panel-background-color) 2px
+
+    .item
+      text-align: left
+      padding: 2px 20px
+      cursor: pointer
+      width: 100%
+
+      &:hover
+        background-color: var(--accent-secondary-hover-color)
+
+.content-page
+    display: flex
+    flex-direction: column
+    gap: 1em
