+++ pages/lex/+Page.ts
@@ -4,10 +4,13 @@ import { apiV2Prefix } from "@macrostrat-web/settings";
 import { useAPIResult } from "@macrostrat/ui-components";
 import h from "./+Page.module.sass";
 import { useData } from "vike-react/useData";
-import { Loading, Footer } from "~/components/general";
+import { Footer } from "~/components/general";
+import { SearchBar } from "~/components/general";
+import { useState } from "react";
 
 export function Page() {
   const { res } = useData();
+  const [ showBody, setShowBody ] = useState(true);
 
   const seen = new Set();
   const stats = res.filter((project) => {
@@ -29,9 +32,8 @@ export function Page() {
   });
 
   return h("div", [
-    h(ContentPage, [
+    h(ContentPage, { className: "content-page" }, [
       h(PageHeader, { title: "Lexicon" }),
-
       h("p", [
         "This is the homepage of Macrostrat's geological lexicons, which are assembled from many data sources including Canada's ",
         h(
@@ -50,42 +52,46 @@ export function Page() {
         h("p.stat", `${formatNumber(measurements)} measurements`),
       ]),
 
-      h("h2", "Dictionaries"),
-      h(
-        LinkCard,
-        { href: "/lex/strat-names", title: "Stratigraphic names" },
-        "Names of rock units, organized hierarchically and concepts that capture relationships between differently-named rock units"
-      ),
+      h(SearchContainer, { setShowBody}),
+      
+      h.if(showBody)('div.body-content', [
+        h("h2", "Dictionaries"),
+        h(
+          LinkCard,
+          { href: "/lex/strat-names", title: "Stratigraphic names" },
+          "Names of rock units, organized hierarchically and concepts that capture relationships between differently-named rock units"
+        ),
 
-      h(
-        LinkCard,
-        { href: "/lex/intervals", title: "Intervals" },
-        "Time intervals"
-      ),
-      h(
-        LinkCard,
-        { href: "/lex/timescales", title: "Timescales" },
-        "Continuous representations of relative geologic time"
-      ),
-      h(
-        LinkCard,
-        { href: "/lex/lithology", title: "Lithologies" },
-        "Names of geologic materials"
-      ),
-      h(
-        LinkCard,
-        { href: "/lex/environments", title: "Environments" },
-        "Depositional environments and formation mechanisms"
-      ),
-      h(
-        LinkCard,
-        { href: "/lex/economics", title: "Economics" },
-        "Economic uses of geologic materials"
-      ),
+        h(
+          LinkCard,
+          { href: "/lex/intervals", title: "Intervals" },
+          "Time intervals"
+        ),
+        h(
+          LinkCard,
+          { href: "/lex/timescales", title: "Timescales" },
+          "Continuous representations of relative geologic time"
+        ),
+        h(
+          LinkCard,
+          { href: "/lex/lithology", title: "Lithologies" },
+          "Names of geologic materials"
+        ),
+        h(
+          LinkCard,
+          { href: "/lex/environments", title: "Environments" },
+          "Depositional environments and formation mechanisms"
+        ),
+        h(
+          LinkCard,
+          { href: "/lex/economics", title: "Economics" },
+          "Economic uses of geologic materials"
+        ),
 
-      h("p", [
-        h("strong", h("a", { href: "/sift" }, "Sift")),
-        ", Macrostrat's legacy lexicon app, is still available for use as it is gradually brought into this new framework.",
+        h("p", [
+          h("strong", h("a", { href: "/sift" }, "Sift")),
+          ", Macrostrat's legacy lexicon app, is still available for use as it is gradually brought into this new framework.",
+        ]),
       ]),
     ]),
     h(Footer),
@@ -94,4 +100,73 @@ export function Page() {
 
 function formatNumber(num) {
   return num.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+}
+
+function SearchContainer({setShowBody}) {
+  const [input, setInput] = useState("");
+  const url = apiV2Prefix + "/defs/autocomplete?query=" + input;
+  const data = useAPIResult(url)?.success?.data || [];
+
+
+  if(data && input.length > 0) {
+    setShowBody(false);
+  } else {
+    setShowBody(true);
+  }
+  
+
+  return h("div.search-container", [
+    h(SearchBar, {
+      placeholder: "Search the geologic lexicon...",
+      onChange: (e) => setInput(e),
+    }),
+    h(SearchResults, { data }),
+  ]);
+}
+
+function SearchResults({ data }) {
+
+  const categories = [
+    "columns",
+    "econs",
+    // "econ_types",
+    // "econ_classes",
+    "environments",
+    // "environment_types",
+    // "environment_classes",
+    "groups",
+    "intervals",
+    "lithologies",
+    // "lithology_types",
+    // "lithology_classes",
+    // "lithology_attributes",
+    "projects",
+    "strat_name_concepts",
+    // "strat_name_orphans",
+    // "structures",
+    // "minerals",
+  ];
+
+  return h("div.search-results", [
+      categories?.map((category) => {
+        const items = data?.[category];
+        if (!items || items?.length === 0) return;
+
+        const link = category === "econs" ?
+          "economics" : 
+          category === "lithologies" ?
+          "lithology" : 
+          category === "strat_name_concepts" ?
+          "strat-name-concepts" : category 
+
+        return h("div.search-category", [
+          h("h3.category", category.charAt(0).toUpperCase() + category.slice(1)),
+          h("ul.items", items?.map((item) => {
+            const { name } = item;
+            return h("li", { key: item.id, className: "item" }, h("a", { href: `/lex/${link}/${item.id}` }, name.charAt(0).toUpperCase() + name.slice(1)));
+          })),
+        ]);
+      }),
+    ])
+
 }
